+++ RELEASE_NOTES.md
@@ -1,6 +1,11 @@
 # sample_search_api release notes
 =========================================
 
+0.0.1
+-----
+* Adding filter_samples function for filtering on a list of given samples
+* RE_ADMIN_TOKEN is a required catalog defined variable for full functionality
+
 0.0.0
 -----
 * Module created by kb-sdk init